refactor: add ansible playbooks
fix: remove unnecessary files created by claude code

fix: generate the inventory.yml in correct place

The code generates an inventory.yml inside of the tofu dir but it needs
to generate this file in the build directory.

This commit updates the main.tf to generate the file correctly.

Additional updates:
- updates the ansible.cfg to reference a .yml file instead of a .ini
file

- Adds an example of the expected inventory.yaml file

fix: Use data to look up the public ssh key

Currently terraform is not configured to add my public ssh key to the
provisioned droplets. This prevents being able to ssh as root.

This commit updates the terraform in order to read this information
from my user account and then add my public key to the account.

fix: Add second inventory reference to ansible.cfg

The ansible.cfg needs two references. One is to the dir that has the
group vars and the second is the the dir containing the autogenerated
inventory.yml